ATLAS-183 Move more size_t to idx_t in atlas::array and field

+++ src/atlas/parallel/HaloExchange.cc
@@ -142,12 +142,12 @@ void execute_halo_exchange( HaloExchange* This, Value field[], int var_strides[]
     // WARNING: Only works if there is only one parallel dimension AND being
     // slowest moving
 
-    array::ArrayShape shape{size_t( This->backdoor.parsize )};
-    for ( size_t j = 0; j < var_rank; ++j )
+    array::ArrayShape shape{ This->backdoor.parsize };
+    for ( int j = 0; j < var_rank; ++j )
         shape.push_back( var_extents[j] );
 
-    array::ArrayStrides strides{size_t( var_extents[0] * var_strides[0] )};
-    for ( size_t j = 0; j < var_rank; ++j )
+    array::ArrayStrides strides{ var_extents[0] * var_strides[0] };
+    for ( int j = 0; j < var_rank; ++j )
         strides.push_back( var_strides[j] );
 
     eckit::SharedPtr<array::Array> arr( array::Array::wrap( field, array::ArraySpec{shape, strides} ) );
